# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Safai Market — Anupurna Traders store management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived]
        - name: lowStockOnly
          in: query
          schema:
            type: boolean
        - name: outOfStockOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID with full details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
  /products/{id}/archive:
    post:
      operationId: archiveProduct
      tags: [products]
      summary: Archive a product (soft delete)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Archived product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}/stock-movements:
    get:
      operationId: getProductStockMovements
      tags: [products]
      summary: Get stock movement history for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stock movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"

  # ─── BILLS ───────────────────────────────────────────────────────────────────
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Record a new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DAILY CLOSINGS ───────────────────────────────────────────────────────────
  /daily-closings:
    get:
      operationId: listDailyClosings
      tags: [dailyClosings]
      summary: List daily closing records
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of daily closings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyClosing"
    post:
      operationId: createDailyClosing
      tags: [dailyClosings]
      summary: Submit daily closing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyClosingInput"
      responses:
        "201":
          description: Daily closing submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyClosing"
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, currentStock, sellPrice, buyPrice, unit, status, categoryId]
      properties:
        id:
          type: integer
        skuCode:
          type: ["string", "null"]
        name:
          type: string
        displayName:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        subcategory:
          type: ["string", "null"]
        unit:
          type: string
        buyPrice:
          type: number
        sellPrice:
          type: number
        wholesalePrice:
          type: ["number", "null"]
        mrp:
          type: ["number", "null"]
        currentStock:
          type: number
        lowStockLimit:
          type: number
        reorderQuantity:
          type: ["number", "null"]
        primarySupplierId:
          type: ["integer", "null"]
        hinglishAliases:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, archived]
        isVariantParent:
          type: boolean
        parentProductId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ProductInput:
      type: object
      required: [name, categoryId, unit, buyPrice, sellPrice]
      properties:
        name:
          type: string
        displayName:
          type: string
        brand:
          type: string
        categoryId:
          type: integer
        subcategory:
          type: string
        unit:
          type: string
        buyPrice:
          type: number
        sellPrice:
          type: number
        wholesalePrice:
          type: number
        mrp:
          type: number
        initialStock:
          type: number
        lowStockLimit:
          type: number
        reorderQuantity:
          type: number
        primarySupplierId:
          type: integer
        hinglishAliases:
          type: string
        parentProductId:
          type: integer
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        displayName:
          type: string
        brand:
          type: string
        categoryId:
          type: integer
        subcategory:
          type: string
        unit:
          type: string
        buyPrice:
          type: number
        sellPrice:
          type: number
        wholesalePrice:
          type: number
        mrp:
          type: number
        lowStockLimit:
          type: number
        reorderQuantity:
          type: number
        primarySupplierId:
          type: integer
        hinglishAliases:
          type: string

    # ─── BILL ─────────────────────────────────────────────────────────────────
...
    StockMovement:
      type: object
      required: [id, productId, productName, movementType, quantity, stockBefore, stockAfter, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        movementType:
          type: string
          enum: [sale, purchase, adjustment, return, damage]
        quantity:
          type: number
        stockBefore:
          type: number
        stockAfter:
          type: number
        referenceId:
          type: ["integer", "null"]
        referenceType:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Expense:
      type: object
      required: [id, description, amount, category, paymentMode, createdAt]
      properties:
        id:
          type: integer
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        paymentMode:
          type: string
          enum: [cash, upi]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [description, amount, category, paymentMode]
      properties:
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        paymentMode:
          type: string
          enum: [cash, upi]
        notes:
          type: string

    # ─── DAILY CLOSING ────────────────────────────────────────────────────────
    DailyClosing:
      type: object
      required: [id, date, expectedCash, actualCash, difference, totalSales, totalExpenses, status, createdAt]
      properties:
        id:
          type: integer
        date:
          type: string
        expectedCash:
          type: number
        actualCash:
          type: number
        difference:
          type: number
        totalSales:
          type: number
        totalExpenses:
          type: number
        totalUdhaarGiven:
          type: number
        totalUpiReceived:
          type: number
        billCount:
          type: integer
        status:
          type: string
          enum: [closed, reopened]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    DailyClosingInput:
      type: object
      required: [date, actualCash]
      properties:
        date:
          type: string
        actualCash:
          type: number
        notes:
          type: string