# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Safai Market — Anupurna Traders store management API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  displayName?: string;
  brand?: string;
  categoryId: number;
  subcategory?: string;
  unit: string;
  buyPrice: number;
  sellPrice: number;
  wholesalePrice?: number;
  mrp?: number;
  initialStock?: number;
  lowStockLimit?: number;
  reorderQuantity?: number;
  primarySupplierId?: number;
  hinglishAliases?: string;
  parentProductId?: number;
}