# lib/api-zod/src/generated/types/stockMovement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Safai Market — Anupurna Traders store management API
 * OpenAPI spec version: 0.1.0
 */
import type { StockMovementMovementType } from './stockMovementMovementType';

export interface StockMovement {
  id: number;
  productId: number;
  productName: string;
  movementType: StockMovementMovementType;
  quantity: number;
  stockBefore: number;
  stockAfter: number;
  /** @nullable */
  referenceId?: number | null;
  /** @nullable */
  referenceType?: string | null;
  /** @nullable */
  reason?: string | null;
  createdAt: string;
}